fix #1094 add escape key handler on LinkChanger
